chore: fix file_size_bytes type + stabilize Alembic + make pre-commit consistent
- Change missions.file_size_bytes from INTEGER to BIGINTEGER to support files > 2GB
- Clean up autogenerated migration to avoid dropping PostGIS Tiger extension tables
- Improve alembic/env.py with include_object to ignore tiger/topology schemas
- Make pre-commit mypy and manual mypy consistent by adding sqlalchemy[mypy] and pydantic[mypy] plugins
- Fix remaining mypy errors (unused ignores, untyped decorators, Geometry annotations)
- Update test conftest.py with proper Generator types and Testcontainers PostGIS setup

This resolves the last blocking lint/CI issues and prevents future autogenerate problems with PostGIS.

+++ src/birdseye/db/models.py
@@ -5,7 +5,7 @@ SQLAlchemy models for birdseye.
 from datetime import datetime
 
 from geoalchemy2 import Geometry
-from sqlalchemy import JSON, Column, DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, Column, DateTime, Float, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import DeclarativeBase, relationship
 
 
@@ -21,7 +21,7 @@ class Mission(Base):
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
 
     original_filename = Column(String(255), nullable=False)
-    file_size_bytes = Column(Integer, nullable=True)
+    file_size_bytes = Column(BigInteger, nullable=True)
     duration_seconds = Column(Float, nullable=True)
 
     # Geospatial
